fix: resolve all clippy warnings for zero-warning compliance
Collapse nested if statements, derive Default for enums, use
is_some_and/is_none_or, remove unnecessary cast/conversion,
use sort_by_key, and simplify io::Error construction.

+++ crates/sp_core/src/resolvers/year_resolver.rs
@@ -37,7 +37,7 @@ impl Resolver for YearResolver {
         for &episode in episodes {
             match episode.published_at() {
                 Some(pub_date) => {
-                    let year = pub_date.year() as i32;
+                    let year = pub_date.year();
                     grouped.entry(year).or_default().push(episode);
                 }
                 None => {
